pyproject.toml
[project]
name = "godzilla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "markdown>=3.10.2",
    "model-bakery>=1.23.3",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
